refactor: use `httpService` instead direct `axios` calling, rename `d.ts` file to `.ts`
