+++ packages/client/src/shared/services/httpService.ts
@@ -1,0 +1,1 @@
+export { default as httpService } from 'axios'
